# compiler: memory usage optimization around br_table